requirements.txt
pandas
numpy
polars
matplotlib
seaborn
joypy
nltk
anthropic
python-dotenv
pyspark
boto3
scikit-learn
